requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
scikit-learn==1.4.2
python-dotenv==1.0.1